requirements-cf.txt
Flask==2.3.3
python-dotenv==1.0.0
gspread==5.11.3
google-auth==2.23.3
Flask-Cors==4.0.0
functions-framework==3.5.0